PP-2877 - splitting out login controller functions into separate files. This is to make it easier to cleave  controller away from login, as it's gonna be the dashboard
