+++ app/controllers/login/login-after-register-controller.js
@@ -1,0 +1,8 @@
+'use strict'
+
+// NPM dependencies
+const passport = require('passport')
+
+module.exports = (req, res, next) => {
+  return passport.authenticate('localDirect', {failureRedirect: '/login'})(req, res, next)
+}
