+++ app/controllers/login/no-access-controller.js
@@ -1,0 +1,5 @@
+'use strict'
+
+module.exports = (req, res) => {
+  res.render('login/noaccess')
+}
