+++ app/controllers/login/send-again-get-controller.js
@@ -1,0 +1,5 @@
+'use strict'
+
+module.exports = (req, res) => {
+  res.render('login/send_otp_again')
+}
