+++ app/controllers/register_user_controller.js
@@ -9,7 +9,7 @@ const errorResponse = response.renderErrorView
 const successResponse = response.response
 const registrationService = require('../services/user_registration_service')
 const paths = require('../paths')
-const loginController = require('./login_controller')
+const loginController = require('./login')
 const validations = require('../utils/registration_validations')
 const shouldProceedWithRegistration = validations.shouldProceedWithRegistration
 const validateRegistrationInputs = validations.validateUserRegistrationInputs
