+++ test/integration/login_controller_test.js
@@ -12,7 +12,7 @@ var sinon = require('sinon')
 
 var paths = require(path.join(__dirname, '/../../app/paths.js'))
 var mockSession = require(path.join(__dirname, '/../test_helpers/mock_session.js'))
-var loginController = require(path.join(__dirname, '/../../app/controllers/login_controller.js'))
+var loginController = require(path.join(__dirname, '/../../app/controllers/login'))
 var mockRes = require('../fixtures/response')
 
 var chaiAsPromised = require('chai-as-promised')
@@ -201,11 +201,11 @@ describe('login get endpoint', function () {
     var req = { 'body': { 'username': '', 'password': '' } }
     var res = mockRes.getStubbedRes()
 
-    testController(loginController.logUserin, req, res)
+    testController(loginController.loginUser, req, res)
     expect(req.flash('error') === 'empty_all')
 
     res.locals = { 'flash': {} }
-    testController(loginController.logInGet, req, res)
+    testController(loginController.loginGet, req, res)
     expect(res.locals.flash === {username: 'You must enter a username', password: 'You must enter a password'})
     done()
   })
@@ -214,11 +214,11 @@ describe('login get endpoint', function () {
     var req = { 'body': { 'username': '', 'password': '123' } }
     var res = mockRes.getStubbedRes()
 
-    testController(loginController.logUserin, req, res)
+    testController(loginController.loginUser, req, res)
     expect(req.flash('error') === 'empty_username')
 
     res.locals = { 'flash': {} }
-    testController(loginController.logInGet, req, res)
+    testController(loginController.loginGet, req, res)
     expect(res.locals.flash === {username: 'You must enter a username'})
     done()
   })
@@ -229,7 +229,7 @@ describe('login post endpoint', function () {
     var req = { 'body': { 'username': '', 'password': '' } }
     var res = mockRes.getStubbedRes()
 
-    testController(loginController.logUserin, req, res)
+    testController(loginController.loginUser, req, res)
     expect(req.flash('error') === 'empty_all')
     done()
   })
@@ -238,7 +238,7 @@ describe('login post endpoint', function () {
     var req = { 'body': { 'username': '', 'password': '123' } }
     var res = mockRes.getStubbedRes()
 
-    testController(loginController.logUserin, req, res)
+    testController(loginController.loginUser, req, res)
     expect(req.flash('error') === 'empty_username')
     done()
   })
